Prevent users from deleting their own accounts
Updated ApplicationErrors to include CannotDeleteSelf error.
Added ICurrentUserService to DeleteUserCommandHandler.
Modified Handle method to check for self-deletion attempts.
Updated using directives in DeleteUserCommandHandler.cs.

+++ src/ProductManagement.Application/Features/Users/Commands/DeleteUser/DeleteUserCommandHandler.cs
@@ -1,4 +1,6 @@
 ﻿using ProductManagement.Application.Abstractions.Messaging;
+using ProductManagement.Application.Errors;
+using ProductManagement.Application.Interfaces.Application;
 using ProductManagement.Application.Interfaces.Infrastructure;
 using ProductManagement.Domain.Shared;
 
@@ -6,23 +8,30 @@ namespace ProductManagement.Application.Features.Users.Commands.DeleteUser;
 
 internal class DeleteUserCommandHandler : ICommandHandler<DeleteUserCommand>
 {
+    private readonly ICurrentUserService _currentUserService;
     private readonly IUserService _userService;
 
-    public DeleteUserCommandHandler(IUserService userService)
+    public DeleteUserCommandHandler(ICurrentUserService currentUserService, IUserService userService)
     {
+        _currentUserService = currentUserService;
         _userService = userService;
     }
 
     public async Task<Result> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
-        // TODO:
-        // - Can't delete self.
-
         var userResult = await _userService.FindByIdAsync(request.UserId, cancellationToken);
         if (userResult.IsFailure)
         {
             return Result.Failure(userResult.Error);
         }
+
+        // Can't update self.
+        var currentUserId = _currentUserService.UserId;
+        if (currentUserId == request.UserId)
+        {
+            return Result.Failure(ApplicationErrors.User.CannotDeleteSelf);
+        }
+
         var deleteResult = await _userService.DeleteAsync(request.UserId, cancellationToken);
         if (deleteResult.IsFailure)
         {
